re-organized react component files
