+++ client/components/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { BrowserRouter, Route, Link } from 'react-router-dom';
 import MainPage from './mainPage.jsx';
-import LogIn from './login.js';
-import SignUp from './signup.js';
+import LogIn from './../user/login.js';
+import SignUp from './../user/signup.js';
 import Directory from './directory.jsx';
 import ProfilePage from './profile.jsx';
 import NewsFeed from './newsFeed.jsx';
