+++ client/components/directoryItem.jsx
