+++ client/components/newsFeed.jsx
