+++ client/main.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter, Route } from 'react-router-dom';
-import App from './App.js';
-import SignUp from './signup.js';
-import LogIn from './login.js';
-import Directory from './directory.jsx';
-import MainPage from './mainPage.jsx';
+import App from './components/App.js';
+import SignUp from './user/signup.js';
+import LogIn from './user/login.js';
+import Directory from './components/directory.jsx';
+import MainPage from './components/mainPage.jsx';
